+++ src/commands/admin/DeleteCostitemCommand.php
@@ -42,6 +42,12 @@ class DeleteCostitemCommand extends \controllerframework\controllers\Command {
         
         /** Check that the page was requested from itself via the POST method. */
         if ($_SERVER["REQUEST_METHOD"] == "POST") {
+            /** Validate CSRF token before processing. */ 
+            if (!$this->validateCsrfToken($request)) { 
+                $request->set('errorcode', 'InvalidCsrfToken');
+                return self::CMD_ERROR;
+            }
+            
             $costitem->delete();
             
             $request->set('forwardqueryparams', ['id' => $costitem->activity_id]);
@@ -50,6 +56,7 @@ class DeleteCostitemCommand extends \controllerframework\controllers\Command {
 
         /** the page was requested via the GET method or the POST method did not return a status. */
         $responses = array();
+        $responses['csrf_token'] = $this->getCsrfToken();
         $responses['costitem'] = $costitem;
         $responses['returnpath'] = 'editActivity?id='. $costitem->activity_id;
